requirements.txt
robotframework~=6.0
pexpect==4.8.0
psutil~=6.0.0